Changed the request format for the server side checking
