+++ routes/vulnCodeSnippet.ts
@@ -85,3 +85,34 @@ exports.challengesWithCodeSnippet = () => async (req, res, next) => {
   const challenges = matches.map(m => m.match.trim().substr(26).trim()).join(' ').split(' ')
   res.json({ challenges })
 }
+
+const getVerdict = (vulnLines: number[], selectedLines: number[]) => {
+  let verdict: boolean = true
+  vulnLines.sort((a, b) => a - b)
+  selectedLines.sort((a, b) => a - b)
+  if (vulnLines.length !== selectedLines.length) {
+    verdict = false
+  }
+  for (let i = 0; i < vulnLines.length; i++) {
+    if (vulnLines[i] !== selectedLines[i]) {
+      verdict = false
+    }
+  }
+
+  return verdict
+}
+
+exports.checkVulnLines = () => (req, res, next) => {
+  const challenge = challenges[req.body.key]
+  const vulnLines: number[] = cache[challenge.key].vulnLines
+  const selectedLines: number[] = req.body.selectedLines
+  if (getVerdict(vulnLines, selectedLines)) {
+    res.status(200).json({
+      verdict: true
+    })
+  } else {
+    res.status(200).json({
+      verdict: false
+    })
+  }
+}
