+++ server.ts
@@ -36,7 +36,6 @@ const profileImageFileUpload = require('./routes/profileImageFileUpload')
 const profileImageUrlUpload = require('./routes/profileImageUrlUpload')
 const redirect = require('./routes/redirect')
 const vulnCodeSnippet = require('./routes/vulnCodeSnippet')
-const vulnLines = require('./routes/vulnLines')
 const angular = require('./routes/angular')
 const easterEgg = require('./routes/easterEgg')
 const premiumReward = require('./routes/premiumReward')
@@ -559,7 +558,7 @@ restoreOverwrittenFilesWithOriginals().then(() => {
   app.post('/profile', updateUserProfile())
 
   /* Routes for vulnerable Lines */
-  app.post('/verdict', vulnLines.checkVulnLines())
+  app.post('/verdict', vulnCodeSnippet.checkVulnLines())
 
   /* Route for vulnerable code snippets */
   app.get('/snippets', vulnCodeSnippet.challengesWithCodeSnippet())
